A first actually working Finder-Viewer...and it is fast! The key is change to avoid un-necessary pre-preparing dynamic pages -- this would surely break down as repo numbers grow, being slow with a few. The rest here is very hacked together before cleanup, a first sight that however works nicely. The Summmary View is real, for the MIT repos. Next steps will be to elaborate Finder Search, including using it per the design, to very necessarily limit the repos downloaded and shown. Once the Vuex component is available, we'll use it to hold the current Finder set so that the Viewer can just use the data rather than query again. Further, the Vuex will allow marking all the chosen repo data so it's not purged, making the Alternatives in the left pane resident during further Finds.
We want of course to limit what the Finder can find to Design repos only. This can and has been done by snooping improper repos, but that is definitely hacky, and there are many other uses that can be made of adding a MetaData.md file to the repo standard. It can carry Overview text, and any number of data elements in the front matter -- such as a 'realRepo: true' item, which allows even skipping the doc-template properly, along with any others.